cambios generales para funcionamiento
