+++ packages/telegram/package.json
@@ -7,7 +7,8 @@
   "type": "module",
   "scripts": {
     "test": "jest",
-    "build": "rimraf dist && rollup --config"
+    "build": "rimraf dist && rollup --config",
+    "local:build": "pnpm run build && npm pack"
   },
   "files": [
     "./dist/"
