+++ packages/telegram/src/index.ts
@@ -1,16 +1,23 @@
 import { ProviderClass, utils } from '@bot-whatsapp/bot'
+import { Vendor } from '@bot-whatsapp/bot/dist/provider/providerClass'
 import { Telegraf } from 'telegraf'
-import { Events, GlobalVendorArgs, MessageCreated } from './@types/types'
+import { Events, GlobalVendorArgs, MessageCreated } from './types'
 
 class TelegramProvider extends ProviderClass {
-  public vendor: Telegraf
+  vendor: Vendor<Telegraf>
+  globalVendorArgs: GlobalVendorArgs
 
-  constructor(public globalVendorArgs: GlobalVendorArgs) {
-    super()
-    this.vendor = new Telegraf(
-      this.globalVendorArgs.token,
-      this.globalVendorArgs?.options || undefined
-    )
+  constructor({ token }: Partial<GlobalVendorArgs>) {
+    super();
+    this.globalVendorArgs = { ...this.globalVendorArgs, token }
+
+    this.vendor = new Telegraf(this.globalVendorArgs.token)
+
+    this.initProvider()
+  }
+
+  private initProvider() {
+
     const listEvents = this.busEvents()
 
     for (const { event, func } of listEvents) {
@@ -20,7 +27,7 @@ class TelegramProvider extends ProviderClass {
 
     this.handleError()
     console.info('[INFO]: Provider loaded')
-    this.vendor.launch(this.globalVendorArgs.launchOptions)
+    this.vendor.launch()
   }
 
   private handleError() {
@@ -152,4 +159,4 @@ class TelegramProvider extends ProviderClass {
   }
 }
 
-export { TelegramProvider } 
+export { TelegramProvider }
